Move to pyproject.toml, remove legacy files, update support
