+++ .github/workflows/publish.yml
@@ -1,4 +1,4 @@
-# This workflows will upload a Python Package using Twine when a release is published
+# This workflow will upload a Python Package using Twine when a release is published
 # For more information see: https://help.github.com/en/actions/language-and-framework-guides/using-python-with-github-actions#publishing-to-package-registries
 
 name: Publish package
@@ -11,30 +11,28 @@ jobs:
     runs-on: ubuntu-latest
 
     steps:
-    - uses: actions/checkout@v2
+    - uses: actions/checkout@v4
+      with:
+        fetch-depth: 0
 
     - name: Set up Python
-      uses: actions/setup-python@v2
+      uses: actions/setup-python@v5
       with:
-        python-version: '3.8'
-        architecture: 'x64'
+        python-version: '3.12'
 
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install setuptools wheel twine
+        pip install ".[build]"
+
     - name: Build source and binary distribution package
-      run: |
-        python setup.py sdist bdist_wheel
-      env:
-        PACKAGE_VERSION: ${{ github.ref }}
+      run: python -m build
 
     - name: Check distribution package
-      run: |
-        twine check dist/*
+      run: twine check dist/*
+
     - name: Publish distribution package
-      run: |
-        twine upload dist/*
+      run: twine upload dist/*
       env:
         TWINE_REPOSITORY: ${{ secrets.PYPI_REPOSITORY }}
         TWINE_USERNAME: ${{ secrets.PYPI_USERNAME }}
