+++ pyproject.toml
@@ -1,8 +1,78 @@
-[tool.isort]
-known_first_party = ["updatable"]
-profile = "black"
-multi_line_output = 3
-line_length = 120
+[build-system]
+requires = ["setuptools>=61.0", "setuptools_scm[toml]>=6.2"]
+build-backend = "setuptools.build_meta"
 
-[tool.black]
-line_length = 120
+[project]
+name = "updatable"
+description = "Finds packages that require updates on a python environment."
+readme = "README.rst"
+license = {text = "MIT"}
+authors = [
+    {name = "Harald Nezbeda", email = "hn@nezhar.com"},
+]
+classifiers = [
+    "Programming Language :: Python",
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
+    "Programming Language :: Python :: 3.14",
+    "Programming Language :: Python :: Implementation :: CPython",
+    "Programming Language :: Python :: Implementation :: PyPy",
+    "Intended Audience :: Developers",
+    "Operating System :: OS Independent",
+    "Natural Language :: English",
+    "Topic :: Utilities",
+]
+requires-python = ">=3.10"
+dependencies = [
+    "httpx",
+    "semantic_version",
+    "packaging",
+]
+dynamic = ["version"]
+
+[project.urls]
+Homepage = "https://github.com/nezhar/updatable"
+Repository = "https://github.com/nezhar/updatable"
+
+[project.scripts]
+updatable = "updatable.console:main"
+
+[project.optional-dependencies]
+test = [
+    "coverage",
+    "respx",
+]
+dev = [
+    "pre-commit",
+    "ruff",
+]
+build = [
+    "build",
+    "twine",
+    "wheel",
+]
+
+[tool.setuptools_scm]
+
+[tool.setuptools.packages.find]
+exclude = ["tests*"]
+
+[tool.ruff]
+line-length = 120
+target-version = "py310"
+
+[tool.ruff.lint]
+select = [
+    "E",      # pycodestyle errors
+    "W",      # pycodestyle warnings
+    "F",      # pyflakes
+    "I",      # isort
+    "UP",     # pyupgrade
+    "COM818", # trailing comma
+]
+
+[tool.ruff.lint.isort]
+known-first-party = ["updatable"]
